+++ main.h
@@ -16,7 +16,7 @@ int _addstr(char *buffer, char s[],int n);
 int _strlen(char *s);
 int print_str(char str[], char *buffer, int len);
 void reverse(char str[]);
-void int_str(int n, char s[]);
+void int_str(long int n, char s[]);
 void uint_str(unsigned int n, char s[], unsigned int base);
 void uint_str2(long int n, char s[],unsigned int base);
 void uint_str3(int n, char s[],unsigned int base);
